Added sentiment.py
Added basic naive classifier, trained on movie reviews and probably not
very good at this. But that's ok. It's pretty easy to play with the
python nltk library.

Something we can do wth this is generate JSON to use in visulations of
Locations and maps and things. That will be fun.